requirements.txt
python-dotenv
langchain
langchain-community
langchain-google-genai
chromadb
firecrawl-py==0.0.13
langchain-chroma
